Update UserDisplay > surveys authored
Surveys are now split into two groups: live and unpublished.
There is currently no conditional to display a message
in the case that authored surveys exist but one section is
empty.

This is part of the steps to separate live surveys
from editable surveys and limit edit access to author
only.