Implement edit and live survey routes
Currently, only two components link to a survey:
1. SurveysContainer, iff the survey is live
2. UserDisplay, both live and editable

These three links have been changed, so that a live
survey is accessed at /surveys/live/:id, and editing
an unpublished survey is at /survey/edit/:id.

At this time, the latter route directs to the existing
SurveyDisplay component; the latter renders a template
LiveSurveyDisplay, which only renders the component
title and the :id param.