Build NewResponse component
This component fetches survey data in 3 parts: the survey,
the associated questions, and the response options.

This required adding a new resource route to fetch
the response options associated with a survey.

Currently only the title of the survey and the counts
of the questions and response options are displayed.