Render responses on UserDisplay
We add a /users/responses route, which returns
all responses authored by the user. The responses
controller is updated to handle this route, which
in addition returns the title of the survey associated
with each response option.

This data is rendered in UserDisplay, similar to the way
in which surveys authored is displayed.

Since now UserDisplay needs to fetch from 3 sources (user show,
surveys index, and responses index), we implement
define a fetch with two params and call it three times
inside useEffect (this strategy was first implemented
in e6963ded80e52457a71325ce6767cbdef42df26b.