Update Header menus and responsiveness
First, tailwind meta is to have the default styling be mobile-
first, then adjust for larger screens.

NB: The adjustment needs to explicitly replace the default,
or both will still take effect.
For example:
-left-[86px] and md:right-0
will both act on medium and larger screens.

This commit instead uses
-right-[86px] md:right-0.

We also update the hamburger button to close the user
action menu on click.