Add Loading and 404 to UsersDisplay
The users controller #show is updated: The before method set_user
now calls find_by rather than find, so no error is thrown, and
we can instead return an informative error message.

We display a simply loading skeleton and, in the case that
the request was of a valid form but there is no such user,
we render instead NotFound.

This still isn't the best way to handle this type of error, but
at least we don't 500 on a bad url.