Correct NewSessionForm
The component was not calling setCurrentUser, nor was the
sessions controller returning the session_token,
so no info about the current user was actually being set properly.

This means that the prod issues with redirecting away from Landing,
Register, and Login had nothing to do with state, and
rather were caused by my mistake in designing the components
and the related controller.