Improve session handling
The App now checks for session to determine routes;
most no-current-user routes render instead login.
Both have a 404 catch-all that is properly a route (no more
yellow warnings!)

To accomodate this, there is now a Profile (in addition to
UserDisplay); the former is for the currentUser,
the latter for viewing other users. In the case of the
latter, the unpublished surveys are not fetched; this
is handled on a controller level by comparing
params[:user_id] to current_user.id.

The wording of UserDisplay is updated to reflect this change.

the initial value of currentUser is now {}; it is important
to check for currentUser.id in conditionals, as !{} is false.

Specifically to the session handling, no component is using
(I think) session storage; in this manner currentUser persists
between tabs.