Build parent Auth component
This will render Login or Register depending on props;
default is Login with empty (override) message.
The source is also defaulted to profile/; these
are overridder as needed by the routes
in App.

Swapping between Login and Register maintains the props,
so that registering (instead of loggin in) will still
redirect to the desired resource.