Improve user not found handling
The initial state of the user object is now null, so that the
loading screen will easily be initially displayed. If
user is not null (which remember includes {}), the
remaining two render options depend on user.id; if it is truthy
then we render the user display, otherwise we render
NotFound with a message User not found, and link message
View all users.