Add initial loading display to SurveysContainer
Note that this change required (as before) to change the initial
state of surveys to null; this means that an earlier seen
Each child in a list should have a unique "key" prop error
was due to the initial render mapping over each element
of {[]}, and [].id, the key value, is undefined.

TODO - build the loading skeleton and no surveys exist displays.